requirements.txt
webvtt_py==0.4.6
googletrans==3.1.0a0
